+++ packages/abjad-web/CHANGELOG.md
@@ -7,6 +7,7 @@
 - favicon
 - info about page's GitHub, npm and email
 - Speed Insights
+- Space and Clear all buttons
 
 ## [0.4.2] 2024-01-20
 ### Changed